requirements.txt
pytest==7.2.0
requests~=2.28.1
voluptuous~=0.13.1
pytest-voluptuous
python-dotenv~=0.21.0
mimesis~=7.0.0
selene~=2.0.0b14
allure-pytest